Split out isRequestingPostsForQueryIgnoringPage
